+++ packages/snippets/src/browser/addPageMarker.js
@@ -1,8 +1,7 @@
 const setElementStyleProperties = require('./setElementStyleProperties')
 
-function addPageMarker([{mask, scale = 1}]) {
+function addPageMarker([{mask, size = 1}]) {
   const colors = {1: 'rgb(0, 0, 0)', 0: 'rgb(255, 255, 255)'}
-  const size = scale >= 1 ? 1 : 2
   const offset = size
 
   const marker = document.createElement('div')
@@ -11,7 +10,7 @@ function addPageMarker([{mask, scale = 1}]) {
   marker.style.setProperty('top', '0', 'important')
   marker.style.setProperty('left', '0', 'important')
   marker.style.setProperty('box-sizing', 'content-box', 'important')
-  marker.style.setProperty('padding', `${offset / scale}px`, 'important')
+  marker.style.setProperty('padding', `${offset}px`, 'important')
   marker.style.setProperty('background', `rgb(128, 128, 128)`, 'important')
   marker.style.setProperty('z-index', '2147483647', 'important')
 
@@ -23,8 +22,8 @@ function addPageMarker([{mask, scale = 1}]) {
     const item = document.createElement('div')
     item.style.setProperty('display', 'block', 'important')
     item.style.setProperty('float', 'left', 'important')
-    item.style.setProperty('width', `${length * (size / scale)}px`, 'important')
-    item.style.setProperty('height', `${size / scale}px`, 'important')
+    item.style.setProperty('width', `${length * size}px`, 'important')
+    item.style.setProperty('height', `${size}px`, 'important')
     item.style.setProperty('background', colors[color], 'important')
     marker.appendChild(item)
   }