+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add method `math.multiplier` method to find a smallest integer multiplier for number to produce an integer result with some allowed error
 
  ## 1.2.12 - 2022/2/9